made scenes for everyone
`Assets/Zaid` contains the screen that holds the right side of the screen.
`Assets/Sima` contains the screen that holds the top-down view of the golf field.
`Assets/Nadeem` contains two scenes, the first is the main menu which has a play, options, leaderboard, credits and exit buttons.
`Assets/Mohammed` contains the scene that hold the bottom left of the screen, the stats of the game: scores for dad and opponent, and sus meter.
`Assets/Lime` contains the scene that hold the commentary window.